Airplane Seating
Program One: Fundamental Concepts

The purpose of this assignment is to reinforce basic C++ skill, including arrays/vectors, input and output, loops, functions, and programming conventions.

For this assignment you are to write a program that assigns passenger seats in an airplane.

The basic requirement are:

Use functions, vectors and exceptions as needed

Prompt the user for the number of rows and seats per row in the plane

Display the initial seating assignment (i.e., no seats have been assigned).  For example:
          image.png

Repeatedly ask the user until they no longer wish to continue or there are no seats available
To enter the row and seat number
Display the updated seating chart
Upon completion display the final seating chart, the number of seats still available, and all of the seats that have been taken in the order they were assigned.  For example:
         image.png