Program Three: Lists
The purpose of this lab is to gain experience using lists.

Scenario: Create a hangman game.  In the game of hangman the player is presented with a list of blanks representing the unknown letters of a hidden word.  The player must guess a letter to see if that letter exists in the word.  If it does, the blank(s) of the corresponding letter(s) in the word are replaced with the letter.  If the letter is not in the word the player is incrementally moved towards "being hung" and the letter is added to guessed letters.  The different states of "being hung" are represented by seven different states: PLATFORM, HEAD, BODY, LEFT_ARM, RIGHT_ARM, LEFT_LEG, RIGHT_LEG.

Requirements:

Create a class for word
Use a list to represent the word to be guess (each node is a letter or an underscore that represents an unguessed letter)
Functions
getword() returns the list 
addLetter(char) replaces the underscore with the char where appropriate
findLetter(char) returns true if the guessed letter is present
solved() returns if the word is guessed
Create a class for game state
Use a list to represent the current game state
Functions
getState() returns the list 
addState() add next state to the list
gameOver() returns true if hung
Create a class for the guessed letters
Use a list to store the guessed letters
Functions
getLetters() returns the list 
addLetter(char) add the to list of guessed letters
Use the main program given below to test your classes
#include <iostream>
#include "word.h"
#include "guessedLetters.h"
#include "gameState.h"
using namespace std;

int main()
{
    Word word("JUGGLE");
    GuessedLetters guessed = GuessedLetters();
    GameState state = GameState();
    char letter;

    do {
       cout << "Game state: " << state.getState() << endl;
       cout << "Guessed letters: " << guessed.getLetters() << endl;
       cout << "Guess this word: " << word.getWord() << endl << endl;
       cout << "Guess a letter: ";
       cin >> letter;
       letter = toupper(letter);
       if (word.findLetter(letter)) {
          word.addLetter(letter);
       }
       else {
          guessed.addLetter(letter);
          state.addState();
       }
    } while ( (!state.gameOver()) && (!word.solved()) );

    if (state.gameOver()) {
       cout << endl << "Game state: " << state.getState() << endl;
       cout << "Sorry you lose"  << endl;
       cout << "The word was: JUGGLE" << endl;
    }
    else {
        cout << endl << "Congratulations, you guessed the word: " <<  word.getWord() << endl;
    }

    return 0;
}