Program Four: Files
The purpose of this assignment is to reinforce your understanding of stacks and queues.

You are to implement a class named SpecialStack.  This class implements a stack that holds integer numbers.  It provides the following basic operations: push, pop, peek, empty, and printStack.  However, it should also support two additional pop operations - popOdd and popEven - they remove the latest added even or odd number from the stack while leaving the rest of the stack unchanged. 

For example, pushing the values 77, 12, 13, 22, 42 would yield the following stack:

image.png

and after a  popOdd the stack would be as follows:

image.png

Requirements

Use the Stack template in the STL
Use exceptions to catch/throw errors
Use the following main program 
Submit the URL to the GitHub directory that contains the .h and the .cpp files
Make sure the protection is set correctly
int main()
{
    SpecialStack one, two, three, four;

    try {
       cout << "Stack one"  << endl;
       one.push(41);
       one.push(66);
       one.push(91);
       one.push(88);
       one.push(33);
       one.popEven();
       one.printStack();
    }
    catch (int e) {
       cout << "Unable to complete operations" << endl << endl;
    }

    try {
       cout << "Stack two" << endl;
       two.push(1);
       two.push(9);
       two.push(6);
       two.push(2);
       two.push(8);
       two.push(6);
       two.push(4);
       two.popOdd();
       two.printStack();
    }
    catch (int e) {
       cout << "Unable to complete operations" << endl << endl;
    }

    try {
        cout << "Stack three" << endl;
        three.popEven();
        three.printStack();
     }
    catch (int e) {
       cout << "Unable to complete operations" << endl << endl;
    }

    try {
       cout << "Stack four" << endl;
       four.pop();
       four.printStack();
    }
    catch (int e) {
       cout << "Unable to complete operations" << endl << endl;
    }
    return 0;
}