Program Five: Graphs
Write a main function that using the program we wrote in class to implement the following weighted directed graph and that does a DFS and a BFS search from node E printing out the paths and the cost of each traversal (assume there is no cost to "backing up").  Then write a function (adding it to the WeightedGraph class) that takes two nodes (source and destination) and determines if there is a path from the source to the destination and if there is a path outputs the path and the length of the path or if there is not one it outputs an appropriate message.  Also, add the appropriate statements to the main program to test this function.